requirements.txt
# Data access
pygbif>=0.6.3
huggingface_hub>=0.20.0

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0
scikit-learn>=1.3.0

# ML - install torch manually first: https://pytorch.org/get-started/locally/
torch>=2.1.0
torchvision>=0.16.0
open-clip-torch>=2.24.0
torchmetrics>=1.3.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
Pillow>=10.0.0
